#6 Add basic module split
