+++ selenide-junit5/build.gradle
@@ -13,6 +13,10 @@ repositories {
 
 dependencies {
     compile('org.junit.jupiter:junit-jupiter:5.5.2')
+    compile('org.junit.platform:junit-platform-launcher:1.6.0')
+    compile project(":koverj-java-commons")
+    compile group: 'com.codeborne', name: 'selenide', version: '5.4.1'
+
 }
 
 test{
