+++ selenide-junit5/src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,0 +1,1 @@
+io.koverj.agent.selenide.junit5.listener.LocatorTestListener
