--- src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,0 @@
-io.koverj.agent.listener.LocatorTestListener
